use cmake variables from OyranosConfig.cmake in app start script
Install Oyranos modules into Synnefo.app . But that needs more work to
cover dependencies as well.

fix the usage of the MacOS/Synnefo as a script to start the MacOS/Synnefo.bin binary
I found yet no other way to specify the started executeable inside app/Contents/MacOS/
directory. The renaming of the binary and the script is not so nice.